refactor: rename files to kebab-case
