+++ src/app/components/avatar-app-intro/typewriter-wrapper.tsx
