--- src/app/components/header/header.module.css
@@ -1,31 +1,0 @@
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0px 1em;
-  border-bottom: 1px solid var(--text-color);
-}
-
-.logo {
-  display: flex;
-  align-items: center;
-}
-
-.logo__text {
-  font-size: 1.6em;
-}
-
-.themeChanger {
-  background-color: var(--dark-grey);
-  color: var(--white);
-  padding: 0.5em 1em;
-  border: none;
-  border-radius: 0.5em;
-  cursor: pointer;
-}
-
-@media (max-width: 360px) {
-  .logo__text {
-    font-size: 1.2em;
-  }
-}
