+++ src/app/page.tsx
@@ -1,6 +1,6 @@
-import AvatarAppIntro from "./components/avatarAppIntro/avatarAppIntro";
+import AvatarAppIntro from "./components/avatar-app-intro/avatar-app-intro";
 import Recipe from "./components/recipe/recipe";
-import SearchBar from "./components/searchBar/searchBar";
+import SearchBar from "./components/search-bar/search-bar";
 import styles from "./page.module.css";
 
 export default function Home() {
